renames folder and shorcut function spaceBar
